third_party/WebKit/LayoutTests/imported/wpt/preload/avoid_delaying_onload_link_preload-expected.txt

This is a testharness.js-based test.
FAIL Makes sure link preload preloaded resources are not delaying onload assert_equals: expected 2 but got 0
Harness: the test ran to completion.
